Fix: makefile now has rules for header changes
If a header was changed, a rebuild for that obj
file was never triggered because the rule for an
obj file depended only on the cpp, not the .h.

Modify the pattern rule for all cpp's so that
we now have:

stat_util.obj : stat_util.cpp stat_util.h

vs. just:

stat_util.obj : stat_util.cpp

Also, main.cpp needed its own rule bc
there is no main.h so the rule for it
doesn't match.

Makefile:
 * create new rule for main.cpp
 * modify obj rules to depend on .h as well
   as on the .cpp.