Implement pythons numpy.var in stat_util
The goal of this commit was to create the same
functionality that numpy.var provides. This
was done using the same arg structs as we used
in implementing the numpy.mean. However,
there are some restrictions, and our
implementation doesn't flatten the array in the
case of no axis. Instead, in the case that
mean recieves a 2d array, an axis must be
provided. We secure this requirement using
std::optional, vs using a default value.

src/main.cpp:
 * Do some small stat tests to ensure the
   fn behaves the same as numpy's var fn.
 * Add in extra arrays, with the future
   goal to use them in lieu of the current
   arrays that have variables as rows and
   the observations as columsn. Ie, storing
   age and weight of 100 ppl would be in
   a 2x100 array, whereas python does it
   as 100 x 2.

src/stat_util.cpp:
 * create the var fn that takes an array_like
 * create the var fn that takes a vector
   of doubles.
 * some fixes and comments to code for mean.

include/stat_util.h:
 * declare the var fn's.