Bugfix and improvement of covariance fn
Bugfix: Was dividing the covariance by the
number of variables vs the number of
observations. This would give the correct
result when the number of observations was 2.

Improvement: Python's numpy.cov has the
optional bias argument, which dividies the
squared errors by N-1 by default. Funnily
enough, this improvement was made by necessity.

In python's example, a 2 var with 3 observations
example was giving a cov matrix of
[[1, -1], [-1, 1]].

In our implementation I was getting this result
when I (erronously) divided by the number
of variables. When I fixed it to divide by the
number of observations, I got the same matrix,
but with 0.6666 and -0.6666 respectively. After
figuring out that this result was correct with
a hand calculation, I went back to the python
example on the docs, and realized that there is
a bias parameter that defaults to dividing by
N - 1 vs n, which is why I was getting correct
answers earlier, when dividing by the number of
variables (2). My implementation now behaves
correctly. For the linear regression function,
it sets the bias to true, so that we are dividing
by n, and not n-1, as required by the algorithm.

include/stat_util.h:
 * change the arg struct for the cov fn from
   the one used in the means to one that has
   an option for bias.

src/stat_util.cpp:
 * implement the bugfixes and bias implementation
   as illustrated above.