Add the rowvar option to covariance calculation
The default is to have the rows be variables, and
the columsn be observations. But, we need to be
able to calculate with colums as the vars, and
the rows as the observations, which is how data
is fed into the LinearRegression predictor obj.

We have the goal of not copying the data, which
is why this is something I have to do, vs. simply
transposing the double array.

This also required a small bugfix to the mean
function, which was dividing by the number of
variables vs the number of observations in the
case when axis == 0.

include/stat_util.h:
 * add the rowvar bool value in the args struct

src/stat_util.cpp:
 * add in rowvar functionality to cov.
 * fix the mean fn as described above.