Step 1 of code reformatting - namespace
instead of putting everything in a block,
we should have each function simply scoped.
This helps us make better use of the 8 spaces
on the left of the screen, vs reserving
all of it to scope a namespace. Plus, it's
now a little clearer because you can see the
namespace in each function's name.

This is part 1, because in the next commit,
we'll remove the tabs so that git has an
easier time looking at the history.

src/stat_util.cpp:
 * resolve namespace in function name, vs in
   a large block.

ie, instead of:

namespace foo {
	void bar(void) {
		cout << "bar";
	}

	void baz(void) {
		cout << "baz";
	}
}

we do:

void foo::bar(void) {
	cout << "bar";
}

void foo::baz(void) {
	cout << "baz";
}