Better conformance to sklearn LinearRegression
Instead of taking a 2d array that has vars
as the rows, and the observations as the cols,
we now take a 2d array where the rows are the
observations and the columns are the vars.

However, to maintain the rest of the function,
we create our own copy of the data and process
it, so that the rest of the code doesn't change,
just the name of the 2d array we operate on.

Plus some minor chnages from stat_util to
stat_util_experimental.

The tests file reflects the change in input
accordingly.

src/linear_model.cpp:
 * change fit fn to take the right 2d array
   type, and do so by creating an internal
   copy.
 * change the score fn to take the right
   style of 2d array as well.
 * NOTE: Predict still takes the old style
   of array, and needs to be fixed.

src/tests.cpp:
 * fix the input arrays to be the right
   dimensions for linear regression fit
   and score tests.
 * NOTE: Predict still takes the old style
   of array, and needs to be fixed.